Prevent possible buffer overflow
During loading of the config file, fscanf reads the class_name and icon
into fixed size buffers. So far it would read strings until a
whitespace was encountered, possibly overflowing the buffer.

This commit limits the number of characters that will be read into the
buffers to be below their size.